Move search functions to search.{h,cpp}
